Moved most of the models into the model project
