+++ QuoteApp.Database/Contact/Contact.cs
