+++ QuoteApp.Database/Invoice/Invoice.cs
@@ -1,12 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Data.Entity;
-using System.Data.Entity.Validation;
 using System.Linq;
-using System.Web;
+using QuoteApp.Database.Work;
 
-namespace QuoteApp.Models
+namespace QuoteApp.Database.Invoice
 {
     public class Invoice
     {
@@ -20,7 +17,7 @@ namespace QuoteApp.Models
 
         [Required]
         public int ContactId { get; set; }
-        public virtual Contact Contact { get; set; }
+        public virtual Contact.Contact Contact { get; set; }
 
         [Required]
         public DateTime InvoiceDate { get; set; }
