+++ QuoteApp.Database/Invoice/InvoiceViewModel.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Web;
 
-namespace QuoteApp.Models
+namespace QuoteApp.Database.Invoice
 {
     public class InvoiceViewModel
     {
