+++ QuoteApp.Database/Quote/QuoteSummary.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace QuoteApp.Models
+﻿namespace QuoteApp.Database.Quote
 {
     public class QuoteSummary
     {
