+++ QuoteApp.Database/Work/PartialWorkViewModel.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿using System.Collections.Generic;
 
-namespace QuoteApp.Models
+namespace QuoteApp.Database.Work
 {
     public class PartialWorkViewModel
     {
