+++ QuoteApp.Database/Work/ScheduleWorkViewModel.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Web;
+using QuoteApp.Database.Quote;
 
-namespace QuoteApp.Models
+namespace QuoteApp.Database.Work
 {
     public class ScheduleWorkViewModel
     {
