+++ QuoteApp.Database/Work/ScheduledWork.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using Newtonsoft.Json.Schema;
+﻿using System.Collections.Generic;
 
-namespace QuoteApp.Models
+namespace QuoteApp.Database.Work
 {
     public class ScheduledWork
     {
