+++ QuoteApp.Database/Work/Work.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
+﻿using System.Collections.Generic;
 
-namespace QuoteApp.Models
+namespace QuoteApp.Database.Work
 {
     public class Work
     {
