+++ QuoteApp.Database/Work/WorkItem.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
-
-namespace QuoteApp.Models
+﻿namespace QuoteApp.Database.Work
 {
     public class WorkItem
     {
