+++ QuoteApp.IServices/IApplicationService.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace QuoteApp.IServices
+namespace QuoteApp.Services
 {
     public interface IApplicationService
     {
